pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "interface-expander"
version = "0.1.0"
description = "A Python library for controlling the Interface Expander device"
authors = [
    { name = "AlmCoding", email = "your.email@example.com" }
]
license = { file = "LICENSE.md" }
readme = "README.md"
dependencies = [
    "pyserial>=3.5"
]

[tool.setuptools]
packages = ["interface_expander"]

[tool.setuptools.package-data]
interface_expander = ["../libraries/*"]  # Ensure external libs are included

[project.urls]
"Homepage" = "https://github.com/AlmCoding/expander-py"
"Repository" = "https://github.com/AlmCoding/expander-py"
